refactor: remove MagicAI package and integrate lead extraction functionality into LeadService
