+++ config/app.php
@@ -213,7 +213,6 @@ return [
         Webkul\Marketing\Providers\MarketingServiceProvider::class,
         Webkul\Installer\Providers\InstallerServiceProvider::class,
         Webkul\Lead\Providers\LeadServiceProvider::class,
-        Webkul\MagicAI\Providers\MagicAIServiceProvider::class,
         Webkul\Product\Providers\ProductServiceProvider::class,
         Webkul\Quote\Providers\QuoteServiceProvider::class,
         Webkul\Tag\Providers\TagServiceProvider::class,
