--- packages/Webkul/MagicAI/src/Facades/MagicAI.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Webkul\MagicAI\Facades;
-
-use Illuminate\Support\Facades\Facade;
-
-class MagicAI extends Facade
-{
-    /**
-     * Get the registered name of the component.
-     *
-     * @return string
-     */
-    protected static function getFacadeAccessor()
-    {
-        return 'magic_ai';
-    }
-}
