--- packages/Webkul/MagicAI/src/Http/helpers.php
@@ -1,13 +1,0 @@
-<?php
-
-if (! function_exists('magic_ai')) {
-    /**
-     * MagicAI helper.
-     *
-     * @return \Webkul\MagicAI\MagicAI
-     */
-    function magic_ai()
-    {
-        return app('magic_ai');
-    }
-}
